channel mapの使い方を間違えていたので、修正

+++ FilterTimeFrameSliceByTrack.h
@@ -95,7 +95,7 @@ protected:
 
    std::map<int, std::string> detectorNameMap;
    std::map<int, std::string> detectorPlaneMap;
-   std::map<int, int> detectorSegmentMap;
+   std::map<int, uint8_t> detectorSegmentMap;
    void DefineDetectorIdMap(); // for reading detector configuration files from AnalyzerT103
 
    bool RegisterDetectorConfig_Geometry();
